requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# Deep learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Image processing
Pillow==10.0.0
opencv-python==4.8.0.74
albumentations==1.3.1

# Data handling
kaggle==1.5.16
tqdm==4.65.0

# Utilities
python-dotenv==1.0.0
tensorboard==2.13.0

# Jupyter and visualization
jupyter==1.0.0
notebook==7.0.0
ipywidgets==8.1.0

# Model evaluation
torchmetrics==1.0.1
wandb==0.15.8